Fix a bug in displaying eye position inside virtual trackbox
Y coordinate of the eyes was calculated incorrectly and so the
vertical eye position was incorrectly drawn in the virtual track box.

In coversion from ADA to psychpy norm we need to do a mirroring of
the y coordinate, because in ADA coordinates 0 is the top vertical
position and 1 is the bottom, which means that moving a point with a
positive value vertically leads to move the point downward.
However in PsychoPy norm coordinate system the top y value is 1 and
the bottom is -1, so moving a point with a positive value vertically
leads to move the point upward (in the oposite direction).